deposito y retiro

+++ src/main/java/com/nttdata/model/Product.java
@@ -1,5 +1,7 @@
 package com.nttdata.model;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -12,6 +14,6 @@ public class Product {
 	public String code;
 	public String name;
 	public String type;
-	public double saldo;
+	public BigDecimal saldo;
 	
 }
